refactor(ontoma): move utils.py to common
