+++ src/literature/method/ontoma/common/utils.py
