+++ src/literature/method/ontoma/schemas/__init__.py
@@ -1,0 +1,3 @@
+"""DataFrame JSON schemas."""
+
+from __future__ import annotations
